requirements.txt
fastapi==0.115.9
python-dotenv==1.1.0
pydantic==2.11.5
sentence-transformers==4.1.0
requests==2.32.3
PyJWT==2.10.1
chromadb==1.0.11
pypdf==5.5.0
nltk==3.9.1
google-genai==1.17.0
